+++ NEWS.md
@@ -1,3 +1,11 @@
+# rteachtools 0.0.3
+
+* Resolved issue with missing diagrams in notebook
+
+# rteachtools 0.0.2
+
+* Deployment also works with included diagrams
+
 # rteachtools 0.0.1
 
 * Added a `NEWS.md` file to track changes to the package.